Migrate `StatusGroupMembership` to `Memberships::Status` with `type` column.
This makes status membership queries faster and cleaner.

Dropping the association to `StatusGroupMembershipInfo` for now. This
has never made it to a full feature.

https://trello.com/c/uizDvLne/1108-statusgroup-als-sti-mit-type

https://trello.com/c/D3onnCh1/4-refactoring-memberships
