+++ app/models/concerns/corporation_caching.rb
@@ -1,7 +1,6 @@
 concern :CorporationCaching do
 
   included do
-    cache :status_group_ids
   end
 
 end
