+++ app/models/concerns/dag_link_types.rb
@@ -6,7 +6,7 @@ concern :DagLinkTypes do
 
   def change_type_according_to_other_attributes
     self.type = "Membership" if ancestor_type == "Group" and descendant_type == "User"
-    self.type = "Memberships::Status" if ancestor_type == "Group" and descendant_type == "User" and descendant.kind_of?(StatusGroup)
+    self.type = "Memberships::Status" if self.type == "Membership" and ancestor.kind_of?(StatusGroup)
   end
 
 end
